Fix bug in opossum exporter
When checking if an external attribution already was created the comparison never returned True as a dictionary was compared with an instance of Attribution.
In order to resolve this conversion to the dict (which is ultimately written in the json) file is done at a later stage. Furthermore, a type annotation was added.

Signed-off-by: Nico Carl <nico.carl@tngtech.com>